Refactor page metadata generation and remove unused buildAlternateLocales functionality. Migrate generatePageMetadata to the _service directory for better organization. Clean up related test files and ensure proper imports in page components.
