+++ src/app/[locale]/(common-layout)/user/[handle]/page/[pageSlug]/_domain/build-alternate-locales.test.ts
